GitHub: New implementation of hidden commits (#309)
* GitHub: New implementation of hidden commits

This shows the number of hidden commits on a separate line,
because something was broken with the old implementation
and this is easier to implement.

The maximum number of commits shown for a push is configurable,
and defaults to 3. When the number of commits is only one
over the limit, the last commit will be shown anyway,
because it would be a waste of a line to say "1 more commit"
when the actual commit can be shown on that line.

* GitHub: Change to push.hidden and fix max_announce_commits

I have no idea how that even worked with "self" instead of "self.plugin"
before. Anyway, added the channel to allow channel-specific config too.