Fix issue with getting page text for display
Something changed in Wikipedia or its plugin where by the existing code for reading data from the page itself is broken.  That said, the fix is to use two slashes before the `p[1]` for the `xpath` call.  This allows data to be properly read from the Wikipedia page.

This was tested in `#limnoria-bots` on IRC as well, with my bot, Archangel (which was spun up for this test only).